# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "댓글 및 사진 댓글 CRUD REST API. 인증은 쿠키 기반 세션으로 처리됩니다.",
        "title": "Comment API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/api/comments": {
            "get": {
                "description": "postSeq에 해당하는 댓글을 트리 구조로 반환합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "게시물 댓글 목록 조회",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "게시물 ID",
                        "name": "postSeq",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/comment.CommentResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CommentSession": []
                    }
                ],
                "description": "게시물에 새 댓글을 작성합니다. COMMENT_SESSION 쿠키가 필요합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "댓글 작성",
                "parameters": [
                    {
                        "description": "댓글 작성 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/comment.CreateCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/comments/user/session/heartbeat": {
            "post": {
                "security": [
                    {
                        "CommentSession": []
                    }
                ],
                "description": "COMMENT_SESSION 쿠키의 TTL을 갱신합니다.",
                "tags": [
                    "comments"
                ],
                "summary": "세션 유지 (하트비트)",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/comments/{commentId}": {
            "put": {
                "security": [
                    {
                        "CommentSession": []
                    },
                    {
                        "LifelogSession": []
                    }
                ],
                "description": "댓글 내용을 수정합니다. 작성자(CommentSession) 또는 관리자(LifelogSession)만 수정 가능합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "댓글 수정",
                "parameters": [
                    {
                        "type": "string",
                        "description": "댓글 ID (ObjectID hex)",
                        "name": "commentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "댓글 수정 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/comment.UpdateCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CommentSession": []
                    },
                    {
                        "LifelogSession": []
                    }
                ],
                "description": "댓글을 소프트 삭제합니다. 작성자(CommentSession) 또는 관리자(LifelogSession)만 삭제 가능합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "댓글 삭제",
                "parameters": [
                    {
                        "type": "string",
                        "description": "댓글 ID (ObjectID hex)",
                        "name": "commentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/comments/{commentId}/replies": {
            "post": {
                "security": [
                    {
                        "CommentSession": []
                    }
                ],
                "description": "특정 댓글에 답글을 작성합니다. 최대 depth=2까지 허용됩니다. COMMENT_SESSION 쿠키가 필요합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "답글 작성",
                "parameters": [
                    {
                        "type": "string",
                        "description": "부모 댓글 ID (ObjectID hex)",
                        "name": "commentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "답글 작성 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/comment.CreateReplyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/photo-comments": {
            "get": {
                "description": "photoSeq에 해당하는 사진 댓글 목록을 반환합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photo-comments"
                ],
                "summary": "사진 댓글 목록 조회",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "사진 ID",
                        "name": "photoSeq",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/photocomment.PhotoCommentResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CommentSession": []
                    }
                ],
                "description": "사진에 새 댓글을 작성합니다. COMMENT_SESSION 쿠키가 필요합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photo-comments"
                ],
                "summary": "사진 댓글 작성",
                "parameters": [
                    {
                        "description": "사진 댓글 작성 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/photocomment.CreatePhotoCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/photo-comments/{commentId}": {
            "put": {
                "security": [
                    {
                        "CommentSession": []
                    },
                    {
                        "LifelogSession": []
                    }
                ],
                "description": "사진 댓글 내용을 수정합니다. 작성자(CommentSession) 또는 관리자(LifelogSession)만 수정 가능합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photo-comments"
                ],
                "summary": "사진 댓글 수정",
                "parameters": [
                    {
                        "type": "string",
                        "description": "댓글 ID (ObjectID hex)",
                        "name": "commentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "사진 댓글 수정 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/photocomment.UpdatePhotoCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CommentSession": []
                    },
                    {
                        "LifelogSession": []
                    }
                ],
                "description": "사진 댓글을 소프트 삭제합니다. 작성자(CommentSession) 또는 관리자(LifelogSession)만 삭제 가능합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photo-comments"
                ],
                "summary": "사진 댓글 삭제",
                "parameters": [
                    {
                        "type": "string",
                        "description": "댓글 ID (ObjectID hex)",
                        "name": "commentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/auth/github": {
            "get": {
                "description": "GitHub OAuth 인증 페이지로 리다이렉트합니다. OAUTH_STATE 쿠키가 발급됩니다.",
                "tags": [
                    "auth"
                ],
                "summary": "GitHub OAuth 로그인 시작",
                "responses": {
                    "302": {
                        "description": "Found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/auth/github/callback": {
            "get": {
                "description": "GitHub에서 받은 code와 state로 토큰 교환 후 COMMENT_SESSION 쿠키를 발급합니다. AUTH_SUCCESS_URL 설정 시 해당 URL로 리다이렉트합니다.",
                "tags": [
                    "auth"
                ],
                "summary": "GitHub OAuth 콜백 처리",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GitHub OAuth authorization code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CSRF state 값",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "302": {
                        "description": "Found"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "CommentSession": []
                    }
                ],
                "description": "COMMENT_SESSION 쿠키를 만료시키고 Redis에서 세션을 삭제합니다. 관리자 세션(LIFELOG_SESSION)은 대상이 아닙니다.",
                "tags": [
                    "auth"
                ],
                "summary": "로그아웃",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "comment.CommentResponse": {
            "type": "object",
            "properties": {
                "authorAvatarUrl": {
                    "type": "string"
                },
                "authorEmail": {
                    "type": "string"
                },
                "authorName": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "depth": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "postSeq": {
                    "type": "integer"
                },
                "replies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/comment.CommentResponse"
                    }
                }
            }
        },
        "comment.CreateCommentRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "postSeq": {
                    "type": "integer"
                }
            }
        },
        "comment.CreateReplyRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
        "comment.UpdateCommentRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
        "photocomment.CreatePhotoCommentRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "photoSeq": {
                    "type": "integer"
                }
            }
        },
        "photocomment.PhotoCommentResponse": {
            "type": "object",
            "properties": {
                "authorAvatarUrl": {
                    "type": "string"
                },
                "authorEmail": {
                    "type": "string"
                },
                "authorName": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "photoSeq": {
                    "type": "integer"
                }
            }
        },
        "photocomment.UpdatePhotoCommentRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "CommentSession": {
            "description": "GitHub SSO로 발급된 일반 사용자 세션 쿠키",
            "type": "apiKey",
            "name": "COMMENT_SESSION",
            "in": "cookie"
        },
        "LifelogSession": {
            "description": "lifelog(Java)에서 발급된 관리자 세션 쿠키 (읽기 전용)",
            "type": "apiKey",
            "name": "LIFELOG_SESSION",
            "in": "cookie"
        }
    }
}